Refactor main into separate files
